Exopse RegexMatcher as part of the SDK.
